@@ -1,35 +1,37 @@
 {
-  "name": "rin-quest",
-  "version": "1.0.0",
-  "description": "Nuxt + Electron",
-  "author": "SalzstangeSalamiBrief",
+  "name": "rinquestv2",
   "private": true,
-  "main": "main.js",
-  "build": {
-    "appId": "",
-    "directories": {
-      "buildResources": "static"
-    }
-  },
+  "version": "1.0.0",
+  "author": "SalzstangeSalamiBrief",
+  "type": "module",
   "scripts": {
-    "dev": "cross-env NODE_ENV=DEV electron .",
-    "build": "nuxt build && electron-builder"
+    "dev": "vite",
+    "build": "tsc && vite build",
+    "lint": "eslint . --ext ts,tsx --report-unused-disable-directives --max-warnings 0",
+    "preview": "vite preview"
   },
   "dependencies": {
-    "nuxt": "^2.11.0"
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0"
   },
   "devDependencies": {
-    "@nuxtjs/tailwindcss": "^1.3.3",
-    "babel-eslint": "^10.1.0",
-    "cross-env": "^7.0.0",
-    "electron": "8.0.2",
-    "electron-builder": "^22.3.2",
-    "electron-devtools-installer": "^2.2.4",
-    "eslint": "^6.1.0",
-    "eslint-config-airbnb-base": "^14.0.0",
-    "eslint-friendly-formatter": "^4.0.1",
-    "eslint-loader": "^3.0.3",
-    "eslint-plugin-import": "^2.18.2",
-    "eslint-plugin-vue": "^6.2.1"
+    "@types/react": "^18.2.43",
+    "@types/react-dom": "^18.2.17",
+    "@typescript-eslint/eslint-plugin": "^6.18.1",
+    "@typescript-eslint/parser": "^6.18.1",
+    "@vitejs/plugin-react": "^4.2.1",
+    "autoprefixer": "^10.4.16",
+    "eslint": "^8.56.0",
+    "eslint-config-prettier": "^9.1.0",
+    "eslint-config-react-app": "^7.0.1",
+    "eslint-plugin-react": "^7.33.2",
+    "eslint-plugin-react-hooks": "^4.6.0",
+    "eslint-plugin-react-refresh": "^0.4.5",
+    "postcss": "^8.4.33",
+    "prettier": "3.2.1",
+    "tailwindcss": "^3.4.1",
+    "typescript": "^5.2.2",
+    "vite": "^5.0.8",
+    "vite-plugin-eslint": "^1.8.1"
   }
 }
